+++ hslua-packaging/test/HsLua/Packaging/FunctionTests.hs
@@ -14,16 +14,16 @@ import Data.Maybe (fromMaybe)
 import Data.Version (makeVersion)
 import HsLua.Core (StackIndex, top)
 import HsLua.Packaging.Convenience
+import HsLua.Packaging.Documentation (getdocumentation)
 import HsLua.Packaging.Function
 import HsLua.Packaging.Types
 import HsLua.Marshalling
   ( forcePeek, peekIntegral, peekRealFloat, peekText
   , pushIntegral, pushRealFloat)
-import Test.Tasty.HsLua ((=:), shouldBeResultOf, shouldHoldForResultOf)
+import Test.Tasty.HsLua ((=:), shouldBeResultOf)
 import Test.Tasty (TestTree, testGroup)
 import Test.Tasty.HUnit ((@=?))
 
-import qualified Data.Text as T
 import qualified HsLua.Core as Lua
 
 -- | Calling Haskell functions from Lua.
@@ -86,19 +86,24 @@ tests = testGroup "Call"
     ]
 
   , testGroup "documentation access"
-    [ "pushDocumentation" =:
-      ("factorial (n)\n" `T.isPrefixOf`) `shouldHoldForResultOf` do
+    [ "pushDocumentedFunction pushes one value" =:
+      1 `shouldBeResultOf` do
+        oldtop <- Lua.gettop
         pushDocumentedFunction (factLuaAtIndex 0)
-        numres <- pushDocumentation top
-        Lua.liftIO $ numres @=? Lua.NumResults 1
-        forcePeek $ peekText top
+        newtop <- Lua.gettop
+        pure (newtop - oldtop)
+
+    , "getdocumentation" =:
+      "factorial" `shouldBeResultOf` do
+        pushDocumentedFunction (factLuaAtIndex 0)
+        Lua.TypeTable <- getdocumentation top
+        Lua.TypeString <- Lua.getfield top "name"
+        forcePeek (peekText top)
 
     , "undocumented value" =:
       Lua.TypeNil `shouldBeResultOf` do
         Lua.pushboolean True
-        numres <- pushDocumentation top
-        Lua.liftIO $ numres @=? Lua.NumResults 1
-        Lua.ltype top
+        getdocumentation top
     ]
 
   , testGroup "helpers"